add module-consistency check
