+++ config/watcher_local_test/kustomization.yaml
@@ -28,6 +28,9 @@ patches:
       value: --kyma-requeue-success-interval=20s
     - op: add
       path: /spec/template/spec/containers/0/args/-
+      value: --manifest-requeue-success-interval=5s
+    - op: add
+      path: /spec/template/spec/containers/0/args/-
       value: --log-level=9
     - op: add
       path: /spec/template/spec/containers/0/args/-
