+++ tests/e2e/ca_certificate_rotation_test.go
