+++ tests/e2e/kyma_deletion_test.go
