+++ tests/e2e/purge_metrics_test.go
