+++ tests/e2e/self_signed_certificate_rotation_test.go
