+++ tests/integration/controller/kyma/manifest_test.go
@@ -332,9 +332,9 @@ var _ = Describe("Test Reconciliation Skip label for Manifest", Ordered, func() 
 	})
 
 	It("When a skip label is added to Manifest", func() {
-		Eventually(AddSkipLabelToManifest, Timeout, Interval).
+		Eventually(ConfigSkipLabelToManifest, Timeout, Interval).
 			WithContext(ctx).
-			WithArguments(controlPlaneClient, kyma.GetName(), kyma.GetNamespace(), module.Name).
+			WithArguments(controlPlaneClient, kyma.GetName(), kyma.GetNamespace(), module.Name, true).
 			Should(Succeed())
 	})
